Major commit, not validated yet. Get rid of the ParticleLayer mechanism. Replaced by a Map<String, Object> in Particle.java that contains particle attributes.  Defined getters and setter to particle attributes. Renamed all the *ParticleLayer.java into *Particle.java and made all their functions static with an IParticle as argument. Updated all the classes that relied on the particle layers.

+++ src/org/ichthyop/stage/AbstractStage.java
@@ -71,7 +71,7 @@ public abstract class AbstractStage extends SimulationManagerAccessor {
     private final String blockKey;
 
     public abstract int getStage(IParticle particle);
-    public abstract int getStage(float value);
+    public abstract int getStage(double value);
 
     AbstractStage(BlockType blockType, String blockKey) {
         this.blockType = blockType;
